Refactor packages to gpstest
