+++ library/src/main/java/com/android/gpstest/library/LocationLabelAndData.kt
@@ -2,7 +2,7 @@ package com.android.library
 
 object LocationLabelAndData {
     // Sample conversation data
-    val statusSample = listOf(
+    val locationLabelAndDataSample = listOf(
         "Lat: 28.0713003 °", "Long: -82.4266223 °", "# Sats: 20/25/25", "Bearing:",
         "PDOP: 0.1 m/s", "H/V DOP: 0.5/0.6", "Speed: 0.0 m/s"
     )
